# pkg/client/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: info@green-ecolution.de
    name: Green Ecolution
    url: https://green-ecolution.de
  description: This is the API for the Green Ecolution Management System.
  license:
    name: AGPL
    url: https://raw.githubusercontent.com/green-ecolution/green-ecolution-management/develop/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Green Space Management API
  version: develop
servers:
- url: /
paths:
...
  /v1/vehicle:
    get:
      description: Get all vehicles
      operationId: get-all-vehicles
      parameters:
      - description: Page
        in: query
        name: page
        schema:
          type: integer
      - description: Limit
        in: query
        name: limit
        schema:
          type: integer
      - description: Vehicle Type
        in: query
        name: type
        schema:
          type: string
      - description: Provider
        in: query
        name: provider
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleList'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get all vehicles
      tags:
      - Vehicle
    post:
      description: Create vehicle
      operationId: create-vehicle
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/VehicleCreate'
        description: Vehicle Create Request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Create vehicle
      tags:
      - Vehicle
      x-codegen-request-body-name: body
  /v1/vehicle/{id}:
    delete:
      description: Delete vehicle
      operationId: delete-vehicle
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          content: {}
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Delete vehicle
      tags:
      - Vehicle
    get:
      description: Get vehicle by ID
      operationId: get-vehicle-by-id
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get vehicle by ID
      tags:
      - Vehicle
    put:
      description: Update vehicle
      operationId: update-vehicle
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/VehicleUpdate'
        description: Vehicle Update Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Update vehicle
      tags:
      - Vehicle
      x-codegen-request-body-name: body
  /v1/vehicle/plate/{plate}:
    get:
      description: Get vehicle by plate
      operationId: get-vehicle-by-plate
      parameters:
      - description: Vehicle plate number
        in: path
        name: plate
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get vehicle by plate
      tags:
      - Vehicle
...
components:
  schemas:
...
    DrivingLicense:
      enum:
      - B
      - BE
      - C
      - CE
      - Unknown
      type: string
      x-enum-varnames:
      - DrivingLicenseB
      - DrivingLicenseBE
      - DrivingLicenseC
      - DrivingLicenseCE
      - DrivingLicenseUnknown
...
    HTTPError:
      example:
        path: path
        code: 7
        method: method
        error: error
      properties:
        code:
          type: integer
        error:
          type: string
        method:
          type: string
        path:
          type: string
      required:
      - code
      - error
      - method
      - path
      type: object
...
    Pagination:
      example:
        next_page: 9
        total_records: 4
        total_pages: 2
        prev_page: 3
        current_page: 7
      properties:
        current_page:
          type: integer
        next_page:
          type: integer
        prev_page:
          type: integer
        total_pages:
          type: integer
        total_records:
          type: integer
      required:
      - current_page
      - next_page
      - prev_page
      - total_pages
      - total_records
      type: object
...
    Vehicle:
      example:
        water_capacity: 5.962133916683182
        length: 1.4658129805029452
        number_plate: number_plate
        created_at: created_at
        description: description
        weight: 5.637376656633329
        type: transporter
        additional_information:
          key: ""
        driving_license: B
        updated_at: updated_at
        provider: provider
        width: 2.3021358869347655
        model: model
        id: 6
        height: 0.8008281904610115
        status: active
      properties:
        additional_information:
          additionalProperties: true
          type: object
        created_at:
          type: string
        description:
          type: string
        driving_license:
          $ref: '#/components/schemas/DrivingLicense'
        height:
          type: number
        id:
          type: integer
        length:
          type: number
        model:
          type: string
        number_plate:
          type: string
        provider:
          type: string
        status:
          $ref: '#/components/schemas/VehicleStatus'
        type:
          $ref: '#/components/schemas/VehicleType'
        updated_at:
          type: string
        water_capacity:
          type: number
        weight:
          type: number
        width:
          type: number
      required:
      - additional_information
      - created_at
      - description
      - driving_license
      - height
      - id
      - length
      - model
      - number_plate
      - provider
      - status
      - type
      - updated_at
      - water_capacity
      - weight
      - width
      type: object
    VehicleCreate:
      example:
        water_capacity: 1.4658129805029452
        length: 6.027456183070403
        number_plate: number_plate
        description: description
        weight: 5.962133916683182
        type: transporter
        additional_information:
          key: ""
        driving_license: B
        provider: provider
        width: 5.637376656633329
        model: model
        height: 0.8008281904610115
        status: active
      properties:
        additional_information:
          additionalProperties: true
          type: object
        description:
          type: string
        driving_license:
          $ref: '#/components/schemas/DrivingLicense'
        height:
          type: number
        length:
          type: number
        model:
          type: string
        number_plate:
          type: string
        provider:
          type: string
        status:
          $ref: '#/components/schemas/VehicleStatus'
        type:
          $ref: '#/components/schemas/VehicleType'
        water_capacity:
          type: number
        weight:
          type: number
        width:
          type: number
      required:
      - description
      - driving_license
      - height
      - length
      - model
      - number_plate
      - status
      - type
      - water_capacity
      - weight
      - width
      type: object
    VehicleList:
      example:
        pagination:
          next_page: 9
          total_records: 4
          total_pages: 2
          prev_page: 3
          current_page: 7
        data:
        - water_capacity: 5.962133916683182
          length: 1.4658129805029452
          number_plate: number_plate
          created_at: created_at
          description: description
          weight: 5.637376656633329
          type: transporter
          additional_information:
            key: ""
          driving_license: B
          updated_at: updated_at
          provider: provider
          width: 2.3021358869347655
          model: model
          id: 6
          height: 0.8008281904610115
          status: active
        - water_capacity: 5.962133916683182
          length: 1.4658129805029452
          number_plate: number_plate
          created_at: created_at
          description: description
          weight: 5.637376656633329
          type: transporter
          additional_information:
            key: ""
          driving_license: B
          updated_at: updated_at
          provider: provider
          width: 2.3021358869347655
          model: model
          id: 6
          height: 0.8008281904610115
          status: active
      properties:
        data:
          items:
            $ref: '#/components/schemas/Vehicle'
          type: array
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
      - data
      type: object
    VehicleStatus:
      enum:
      - active
      - available
      - not available
      - unknown
      type: string
      x-enum-varnames:
      - VehicleStatusActive
      - VehicleStatusAvailable
      - VehicleStatusNotAvailable
      - VehicleStatusUnknown
    VehicleType:
      enum:
      - transporter
      - trailer
      - unknown
      type: string
      x-enum-varnames:
      - VehicleTypeTransporter
      - VehicleTypeTrailer
      - VehicleTypeUnknown
    VehicleUpdate:
      example:
        water_capacity: 1.4658129805029452
        length: 6.027456183070403
        number_plate: number_plate
        description: description
        weight: 5.962133916683182
        type: transporter
        additional_information:
          key: ""
        driving_license: B
        provider: provider
        width: 5.637376656633329
        model: model
        height: 0.8008281904610115
        status: active
      properties:
        additional_information:
          additionalProperties: true
          type: object
        description:
          type: string
        driving_license:
          $ref: '#/components/schemas/DrivingLicense'
        height:
          type: number
        length:
          type: number
        model:
          type: string
        number_plate:
          type: string
        provider:
          type: string
        status:
          $ref: '#/components/schemas/VehicleStatus'
        type:
          $ref: '#/components/schemas/VehicleType'
        water_capacity:
          type: number
        weight:
          type: number
        width:
          type: number
      required:
      - description
      - driving_license
      - height
      - length
      - model
      - number_plate
      - status
      - type
      - water_capacity
      - weight
      - width
      type: object
...
  securitySchemes:
    Keycloak:
      flows:
        authorizationCode:
          authorizationUrl: https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/auth
          scopes: {}
          tokenUrl: https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/token
      type: oauth2
...